Added interaction with form fields
