--- src/components/Button/Button.module.css
@@ -1,13 +1,0 @@
-.defaultButton {
-  padding: 14px 40px;
-  border-radius: 30px;
-  border: none;
-  background-color: var(--color-lightgreen);
-
-  /*content styles*/
-  color: var(--color-darkblue);
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 1.28;
-  letter-spacing: -0.14px;
-}
