+++ src/components/WaterModal/WaterModal.module.css
@@ -7,3 +7,11 @@
 
   margin-bottom: 24px;
 }
+
+@media screen and (min-width: 768px) {
+  .title {
+    font-size: 28px;
+    line-height: 1.14;
+    letter-spacing: -0.28px;
+  }
+}
