+++ src/helpers/getCurrentTime.js
@@ -1,0 +1,13 @@
+export const getCurrentTime = () => {
+  const date = new Date();
+  const hours = date.getHours();
+  const minutes = date.getMinutes();
+
+  // Форматуємо для відображення у форматі hh:mm
+  const formattedTime =
+    hours.toString().padStart(2, '0') +
+    ':' +
+    minutes.toString().padStart(2, '0');
+
+  return formattedTime;
+};
